repo ready

+++ app/src/main/java/com/example/myweather/data/implementation/WeatherRepositoryImpl.kt
@@ -1,7 +1,10 @@
 package com.example.myweather.data.implementation
 
 import android.app.Application
+import android.provider.MediaStore.Audio.Media
+import android.util.Log
 import androidx.lifecycle.LiveData
+import androidx.lifecycle.MediatorLiveData
 import androidx.lifecycle.Transformations
 import com.example.myweather.data.api.ApiFactory
 import com.example.myweather.data.db.AppDatabase
@@ -17,41 +20,60 @@ class WeatherRepositoryImpl(application: Application) : WeatherRepository {
     private val apiService = ApiFactory.apiService
 
     override fun getTodayHourlyForecast(): LiveData<List<HourlyForecast>> =
-        Transformations.map(
-            forecastDao.getForecastForToday()
-        ) {
-            it?.map { item -> item.toEntity() }
+        MediatorLiveData<List<HourlyForecast>>().apply {
+            addSource(forecastDao.getForecastForToday()) {
+                if (it != null) {
+                    value = it.map { item -> item.toEntity() }
+                }
+            }
         }
 
     override fun getCurrentForecast(): LiveData<CurrentConditions> =
-        Transformations.map(
-            forecastDao.getCurrentCondition()
-        ) {
-            it?.toEntity()
+        MediatorLiveData<CurrentConditions>().apply {
+            addSource(forecastDao.getCurrentCondition()) {
+                if (it !=null) {
+                    value = it.toEntity()
+                }
+            }
         }
 
     override fun getWeeklyForecast(): LiveData<List<DailyForecast>> =
-        Transformations.map(
-            forecastDao.getForecastForWeek()
-        ) {
-            it?.map { item -> item.toEntity() }
+        MediatorLiveData<List<DailyForecast>>().apply {
+            addSource(forecastDao.getForecastForWeek()) {
+                if (it != null) {
+                    value = it.map { item -> item.toEntity() }
+                }
+            }
         }
 
     override suspend fun loadData(cityName: String) {
-        val weeklyResponse = apiService.getWeeklyForecast(cityName)
-        val todayResponse = apiService.getCurrentAndHourlyForecast(cityName)
 
-        val currentConditions = todayResponse.currentConditions?.toDbModel(cityName)
-            ?: throw RuntimeException("Current conditions requested but not exists.")
+        try {
+            val todayResponse = apiService.getCurrentAndHourlyForecast(cityName)
+            if (todayResponse.isSuccessful) {
+                val currentConditions = todayResponse.body()?.currentConditions?.toDbModel(cityName)
+                    ?: throw RuntimeException("Current conditions requested but not exists.")
 
-        val hourlyForecast = todayResponse.days?.get(0)?.hours?.map { it.toDbModel() }
-            ?: throw RuntimeException("Hourly forecast requested but not exists.")
+                forecastDao.removeAndInsertCurrentCondition(currentConditions)
 
-        val weeklyForecast = weeklyResponse.days?.map { it.toDbModel() }
-            ?: throw RuntimeException("Weekly forecast requested but not exists.")
+                val hourlyForecast = todayResponse.body()?.days?.get(0)?.hours?.map { it.toDbModel() }
+                    ?: throw RuntimeException("Hourly forecast requested but not exists.")
 
-        forecastDao.removeAndInsertCurrentCondition(currentConditions)
-        forecastDao.removeAndInsertForecastForWeek(weeklyForecast)
-        forecastDao.removeAndInsertForecastForToday(hourlyForecast)
+                forecastDao.removeAndInsertForecastForToday(hourlyForecast)
+            }
+        } catch (e: Exception){
+            Log.e("REPO_ERROR", e.message ?: "no message")
+        }
+
+        try {
+            val weeklyResponse = apiService.getWeeklyForecast(cityName)
+            if (weeklyResponse.isSuccessful) {
+                val weeklyForecast = weeklyResponse.body()?.days?.map { it.toDbModel() }
+                    ?: throw RuntimeException("Weekly forecast requested but not exists.")
+                forecastDao.removeAndInsertForecastForWeek(weeklyForecast)
+            }
+        } catch (e: Exception) {
+            Log.e("REPO_ERROR", e.message ?: "no message")
+        }
     }
 }
